Corrected data types

+++ sgl_draw.h
@@ -1,4 +1,4 @@
 
 void SGL_DRAW_hline(SGL_display *display,
-                    uint8_t x0, uint8_t x1,
-                    uint8_t y, uint16_t color);
+                    uint16_t x0, uint16_t x1,
+                    uint16_t y, uint16_t color);
